+++ src/eg3dmath.c
@@ -19,11 +19,11 @@
 /* sin/cos via table lookup + linear interpolation.
  * Mirrors sineLookup: idx = angle>>8, frac = angle&0xFF, then
  *   result = lut[idx] + round((lut[idx+1] - lut[idx]) * frac / 256). */
-static int sineInterp(unsigned angle) {
-    int idx = (angle >> 8) & 0xFF;
-    int frac = angle & 0xFF;
-    int v0 = g_angleLut[idx];
-    int v1 = g_angleLut[idx + 1];
+static int16 sineInterp(uint16 angle) {
+    int16 idx = (angle >> 8) & 0xFF;
+    int16 frac = angle & 0xFF;
+    int16 v0 = g_angleLut[idx];
+    int16 v1 = g_angleLut[idx + 1];
     long step = (long)(v1 - v0) * frac;
     return v0 + (int)((step + 0x80) >> 8);
 }
@@ -34,7 +34,7 @@ int sine(int angle) {
 
 int cosine(int angle) {
     /* cos(x) = sin(x + 90deg); a quarter turn is 0x4000 in 16-bit angle space. */
-    return sineInterp((unsigned)angle + 0x4000);
+    return sineInterp((uint16)angle + 0x4000);
 }
 
 /* Q15-style fixed multiply: returns round((a*b) >> 15).
