+++ src/endbrf.c
@@ -155,7 +155,7 @@ open_dbicons:
     if (commData->trainingFlag == 0) {
         gameData->hallOfFameEligible = missionScore;
 
-        if ((unsigned long)gameData->lastScore < (uint32)missionScore) {
+        if ((uint32)gameData->lastScore < (uint32)missionScore) {
             gameData->lastScore = missionScore;
         }
